Move JSON tests under a package
