+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/01-arrayAtRoot-sameProperties.json
