+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/02-arrayAtRoot-propertiesMissing.json
