+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/03-arrayAtRoot-varyingProperties.json
