+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/04-arrayAtRoot-differentTypes.json
