+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/10-arrayAtRoot-nestedObject.json
