+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/11-arrayAtRoot-nestedArray.json
