+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/12-arrayInObject-sameProperties.json
